+++ src/features/filters/constants.ts
@@ -1,7 +1,7 @@
 import { CommunityBuildFilterProps } from './types'
 
 export const DEFAULT_COMMUNITY_BUILD_FILTERS: CommunityBuildFilterProps = {
-  archtypes: [],
+  archetypes: [],
   longGun: 'All',
   handGun: 'All',
   melee: 'All',