+++ src/features/filters/queries/segments/limitByArchtypes.ts
@@ -1,27 +1,27 @@
 import { remnantItems } from '@/features/items/data'
 import { Prisma } from '@prisma/client'
 
-export function limitByArchtypesSegment(archtypeIds: string[]) {
-  return archtypeIds.length === 0
+export function limitByArchetypesSegment(archetypeIds: string[]) {
+  return archetypeIds.length === 0
     ? Prisma.empty
     : Prisma.sql`AND (
 SELECT COUNT(*)
 FROM BuildItems
 WHERE BuildItems.buildId = Build.id
-AND BuildItems.itemId IN (${Prisma.join(archtypeIds)})
-) = ${archtypeIds.length}`
+AND BuildItems.itemId IN (${Prisma.join(archetypeIds)})
+) = ${archetypeIds.length}`
 }
 
-export function archtypeFiltersToIds({
-  archtypes,
+export function archetypeFiltersToIds({
+  archetypes,
 }: {
-  archtypes: string[]
+  archetypes: string[]
 }): string[] {
   const archtypeIds: string[] = []
 
-  for (const archtype of archtypes) {
+  for (const archetype of archetypes) {
     const item = remnantItems.find(
-      (item) => item.name.toLowerCase() === archtype,
+      (item) => item.name.toLowerCase() === archetype,
     )?.id
     if (item) archtypeIds.push(item)
   }